# Support all string value for key