--- a/.travis.yml
+++ b/.travis.yml
@@ -2,7 +2,10 @@ language: ruby
 rvm:
   - 2.1.8
   - 2.2.4
-  - 2.3.0
+  - 2.3.7
+  - 2.4.6
+  - 2.5.5
+  - 2.6.3
 before_install: gem install bundler -v 1.10.6
 notifications:
   email: false