Fixed #8

$ bundle exec rgot -v --bench . --cpu=1 --thread=1

before:
benchmark_new_instance	480000	2133 ns/op
benchmark_struct_new_instance	7200000	139 ns/op
benchmark_from_hash	86400	11735 ns/op
benchmark_new	86400	13214 ns/op
benchmark_struct_new	72000	15985 ns/op

after:
benchmark_new_instance	864000	1314 ns/op
benchmark_struct_new_instance	7200000	142 ns/op
benchmark_from_hash	103680	10546 ns/op
benchmark_new	86400	12940 ns/op
benchmark_struct_new	72000	15447 ns/op